first_commit
✔ Project Overview:
Analyzing FC 24 data with SQL.
The data contains with three tables players, teams and coaches.
The data gives information about team's nationalities, leagues, league level, players, stadiums, rating, coaches and players' nationality, teams, age, rate and wages.

✔ Steps:
1- download the data from Kaggle
2- Import the data to Excel 
3- Cleaning data in Excel:
- remove unnecessary columns
- Filtering data
4- Import data to SQL Server
5- Analyzing data

✔ Questions asked:
Q1 What is the preferred foot ratio?
Q2 Who are team managers? 
Q3 Who are the club managers in the top 5 leagues?
Q4 What is the number and percentage of local and foreign coaches in the top 5 leagues?
Q5 What is the highest rated team in each league?
Q6 Who are the top-rated players in each team?
Q7 Who are the lowest rated players on each team?
Q8 What is the age distribution of players?
Q9 Who are the ten highest-paid players in the top 5 leagues?
Q10 What are the ten highest paid clubs in the top 5 leagues?
Q11 Who are the highest rated players by position?
Q12 Who are the top-rated players in each country with their teams and league?
Q13 What is the proportion of local and foreign players in each league?
Q14 What is the average age of teams in the top 5 leagues?
Q15 What is the top 5 leagues players average age?
Q16 What are the ages of the coaches?
Q17 All players for each team in Serie A
Q18 What are the teams rating and coaches?
Q19 What is the number of Italian & not Italian players in each team?
Q20 Who is the top paid player in each team?